refactor(affiliate): return full affiliate rows; consolidate swaps tests
- getAffiliates: rename from listAffiliates and return full Affiliate[] with an
  honest return type (endpoint is behind the shared service api key; callers may
  want wallet/receive addresses)
- fold the one useful affiliate-swaps test (fee-split enrichment: happy path +
  unpriceable/no-verified-fee edge) into affiliate.service.test.ts as its own
  section; drop the standalone file and the redundant/tautological cases
